WIP(Solver):add TPDv to solve DarcyForchhrimer Model

+++ fealpy/mesh/edge_mesh.py
@@ -96,7 +96,7 @@ class EdgeMesh(SimplexMesh, Plotable):
 
     Examples
     >>> node = bm.tensor([[0, 0], [1, 0], [2, 0]], dtype=bm.float64)
-    >>> cell = bm.tensor([[0, 1], [1, 2]], dtype=bm.int_)
+    >>> cell = bm.tensor([[0, 1], [1, 2]], dtype=bm.int32)
     >>> mesh = EdgeMesh(node, cell)
     >>> print(mesh.cell_length())
     [1.0, 1.0]
@@ -498,11 +498,11 @@ class EdgeMesh(SimplexMesh, Plotable):
             [1, 3], [1, 4], [0, 2], [0, 5], [2, 5],
             [4, 3], [2, 3], [4, 5], [2, 9], [6, 5], 
             [8, 3], [7, 4], [6, 3], [2, 7], [9, 4],
-            [8, 5], [9, 5], [2, 6], [7, 3], [8, 4]], dtype=bm.int_)
+            [8, 5], [9, 5], [2, 6], [7, 3], [8, 4]], dtype=bm.int32)
         mesh = cls(node, cell)
 
-        mesh.meshdata['disp_bc'] = (bm.tensor([6, 7, 8, 9], dtype=bm.int_), bm.zeros(3))
-        mesh.meshdata['force_bc'] = (bm.tensor([0, 1], dtype=bm.int_), bm.tensor([0, 900, 0]))
+        mesh.meshdata['disp_bc'] = (bm.tensor([6, 7, 8, 9], dtype=bm.int32), bm.zeros(3))
+        mesh.meshdata['force_bc'] = (bm.tensor([0, 1], dtype=bm.int32), bm.tensor([0, 900, 0]))
 
         return mesh 
     
@@ -515,12 +515,12 @@ class EdgeMesh(SimplexMesh, Plotable):
             [400, 300], [0, 300]], dtype=bm.float64)
         cell = bm.tensor([
             [0, 1], [2, 1], 
-            [0, 2], [3, 2]], dtype=bm.int_)
+            [0, 2], [3, 2]], dtype=bm.int32)
         mesh = cls(node, cell)
 
         # 按分量处理自由度索引
-        mesh.meshdata['disp_bc'] = (bm.tensor([0, 1, 3, 6, 7], dtype=bm.int_), bm.zeros(1))
-        mesh.meshdata['force_bc'] = (bm.tensor([1, 2], dtype=bm.int_), 
+        mesh.meshdata['disp_bc'] = (bm.tensor([0, 1, 3, 6, 7], dtype=bm.int32), bm.zeros(1))
+        mesh.meshdata['force_bc'] = (bm.tensor([1, 2], dtype=bm.int32), 
                                      bm.tensor([[2e4, 0], [0, -2.5e4]], dtype=bm.float64))
 
         return mesh 
@@ -534,11 +534,11 @@ class EdgeMesh(SimplexMesh, Plotable):
             [0, 36], [36, 36], [72, 36]], dtype=bm.float64)
         cell = bm.tensor([
             [0, 1], [1, 2], [2, 3],
-            [1, 3], [1, 4], [3, 4]], dtype=bm.int_)
+            [1, 3], [1, 4], [3, 4]], dtype=bm.int32)
         mesh = cls(node, cell)
 
-        mesh.meshdata['disp_bc'] = (bm.tensor([0, 2], dtype=bm.int_), bm.zeros(2))
-        mesh.meshdata['force_bc'] = (bm.tensor([3, 4], dtype=bm.int_), bm.tensor([[0, -500], [0, -500]]))
+        mesh.meshdata['disp_bc'] = (bm.tensor([0, 2], dtype=bm.int32), bm.zeros(2))
+        mesh.meshdata['force_bc'] = (bm.tensor([3, 4], dtype=bm.int32), bm.tensor([[0, -500], [0, -500]]))
 
         return mesh 
     
@@ -551,11 +551,11 @@ class EdgeMesh(SimplexMesh, Plotable):
             [0, 0, -36], [0, 72, 0]], dtype=bm.float64)
         cell = bm.tensor([
             [0, 1], [0, 2], [0, 3],
-            [1, 2], [1, 3], [2, 3]], dtype=bm.int_)
+            [1, 2], [1, 3], [2, 3]], dtype=bm.int32)
         mesh = cls(node, cell)
 
-        mesh.meshdata['disp_bc'] = (bm.tensor([6, 7, 8, 9], dtype=bm.int_), bm.zeros(3))
-        mesh.meshdata['force_bc'] = (bm.tensor([0, 1], dtype=bm.int_), bm.tensor([0, 900, 0]))
+        mesh.meshdata['disp_bc'] = (bm.tensor([6, 7, 8, 9], dtype=bm.int32), bm.zeros(3))
+        mesh.meshdata['force_bc'] = (bm.tensor([0, 1], dtype=bm.int32), bm.tensor([0, 900, 0]))
 
         return mesh 
     
@@ -566,12 +566,12 @@ class EdgeMesh(SimplexMesh, Plotable):
         node = bm.tensor([
             [0], [5], [7.5]], dtype=bm.float64)
         cell = bm.tensor([
-            [0, 1], [1, 2]], dtype=bm.int_)
+            [0, 1], [1, 2]], dtype=bm.int32)
         mesh = cls(node, cell)
 
-        mesh.meshdata['disp_bc'] = (bm.tensor([0, 1], dtype = bm.int_), bm.zeros(2))
-        mesh.meshdata['force_bc'] = (bm.tensor([0, 1, 2], dtype = bm.int_), 
-                                     bm.tensor([[-62500, -52083], [-93750, 39062], [-31250, 13021]], dtype = bm.int_))
+        mesh.meshdata['disp_bc'] = (bm.tensor([0, 1], dtype = bm.int32), bm.zeros(2))
+        mesh.meshdata['force_bc'] = (bm.tensor([0, 1, 2], dtype = bm.int32), 
+                                     bm.tensor([[-62500, -52083], [-93750, 39062], [-31250, 13021]], dtype = bm.int32))
 
         return mesh 
     
@@ -583,11 +583,11 @@ class EdgeMesh(SimplexMesh, Plotable):
             [0, 0.96], [1.44, 0.96], 
             [0, 0], [1.44, 0]], dtype=bm.float64)
         cell = bm.tensor([
-            [0, 1], [2, 0], [3, 1]], dtype=bm.int_)
+            [0, 1], [2, 0], [3, 1]], dtype=bm.int32)
         mesh = cls(node, cell)
 
-        mesh.meshdata['disp_bc'] = (bm.tensor([2, 3], dtype=bm.int_), bm.zeros(3))
-        mesh.meshdata['force_bc'] = (bm.tensor([0, 1], dtype=bm.int_), 
+        mesh.meshdata['disp_bc'] = (bm.tensor([2, 3], dtype=bm.int32), bm.zeros(3))
+        mesh.meshdata['force_bc'] = (bm.tensor([0, 1], dtype=bm.int32), 
                                      bm.tensor([[3000, -3000, -720], 
                                                [0, -3000, 720]], dtype=bm.float64))
